pyproject.toml
[tool.poetry]
name = "drf-starter"
version = "0.1.0"
description = "Starter Template For Django Rest Framework"
authors = ["MahmudulHassan5809 <mahmudul.hassan240@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
pyjwt = "^2.8.0"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"
celery = "^5.3.4"
requests = "^2.31.0"
django-filter = "^23.2"
redis = "^5.0.0"
drf-yasg = "^1.21.7"
boto3 = "^1.28.44"
sendgrid = "^6.10.0"
gunicorn = "^21.2.0"
pika = "^1.3.2"
pymongo = "^4.5.0"
xlsxwriter = "^3.2.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
firebase-admin = "^6.5.0"
django-environ = "^0.12.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
faker = "^19.6.0"
pre-commit = "^3.4.0"
isort = "^5.12.0"
black = "^23.9.0"
mypy = "^1.5.1"
django-stubs = "^4.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
